api/requirements.txt
asgiref==3.2.7
astroid==2.2.5
autopep8==1.5.1
colorama==0.4.1
Django==3.0.5
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.0
httplib2==0.17.2
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.1
typed-ast==1.4.0
wrapt==1.11.2
youtrack==0.1.65
